Update faceting to use ReactiveSearch MultiDropdownList instead of MultiList.  This creates a consistent UI for filtering between Search Results page and the Collection page.  Also updates look and feel of the faceting, to make a white sidebar instead of re-purposing the NU light purple sidebar for usage other than its design intent.
