+++ src/styles/sass/nulib-collections/_nuwebcomm-overrides.scss
@@ -16,7 +16,6 @@
 // Override form inputs only being 50% wide
 .web-form {
   input.rs-search-input {
-    width: 100%;
     padding-left: 30px;
   }
 }
